requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
mysql-connector-python
pymongo
fastapi[standard]
joblib
